fix: use double for avgHeartRate

+++ metric-dto/src/main/java/com/stride/tracking/metric/dto/report/response/ActivityDetailReport.java
@@ -16,5 +16,5 @@ public class ActivityDetailReport {
     private long elevationGain;
     private long distance;
     private int calories;
-    private int avgHeartRate;
+    private double avgHeartRate;
 }
